Backend/requirements.txt
pymongo==4.6.1
python-dotenv==1.0.1
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
motor==3.3.2
openai==1.12.0
passlib[bcrypt]==1.7.4
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0 